+++ src/lib/classes/ErrorHandler.js
